add method to get fxerr value for StatusError code
Need to get StatusError.Code (uint32) converted to fxerr type so you can
compare it to public (exported) ErrSSHFx[..] error types. This adds a
new method, FxCode(), to the StatusError objects to handle this case.

Did this vs trying to change types of existing fields to not break
backwards compatibility.